backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.1
httpx==0.27.2
msal==1.28.0
pydantic==2.9.2
pydantic-settings==2.5.2
azure-identity==1.17.1
azure-search-documents==11.5.1
azure-storage-blob==12.20.0
azure-keyvault-secrets==4.8.0
azure-ai-documentintelligence==1.0.0b3
azure-monitor-opentelemetry-exporter==1.0.0b15
# Align OTel versions with azure exporter requirement (1.19.x)
opentelemetry-sdk==1.19.0
opentelemetry-instrumentation-fastapi==0.40b0
opentelemetry-instrumentation-httpx==0.40b0
opentelemetry-exporter-otlp==1.19.0
opentelemetry-instrumentation-logging==0.40b0
python-dotenv==1.0.1
orjson==3.10.7
python-multipart==0.0.9
beautifulsoup4==4.12.3
markdownify==0.11.6
redis==5.0.3
aiocache==0.12.2
mypy==1.11.2
ruff==0.6.9
black==24.10.0
pytest==8.3.3
pytest-asyncio==0.24.0
aiohttp==3.13.3